Changes: do not encode modification time of ccs file in binary file name
Details:
So far the binary form of the css files has the modification date
of the css file encoded in its filename. If a new version of a css
file had been installed the old one has not been cleaned up.
With this patch the date is encoded in the file itself, so a new
binary file just overwrites the old one.