Changes: Simple kinetics added to slider view
RevBy: Dezso
Details: Added simple kinetic behavior to the sliderview on clicking
the rail: instead of the thumb warping to the value, there's a quick
slide to the value with an easing curve. Also added a timer to keep
the handle from raising and the press state graphics from flickering
into view for a split second if you do a quick press-and-release on
the slider.