Changes: rework widgetsgallery timedemo
RevBy: Kuisma Salonen

Details:
So far the timedemo has been a set of fixed benchmarks.
This patch allows it to add a arbitrary set of benchmarks
to every page in widgetsgallery. Different pages can have
a different set of benchmarks.
So far the only implemented benchmarks are static pages
shown in 0 and 90 degrees.
Benchmarks for the appear or rotation phase of pages are
not written yet, but can be easily added.
Right now we do not have an animation for appearing pages
anyway.