Changes: Assign sequence numbers to the theme daemon IPC packets
RevBy: Trust Me

Details:
* duithemedaemon/duithemedaemonserver.{cpp,h}
(DuiThemeDaemonServer::QueueItem): Introduce new struct to replace the
QPair<> items of pixmapRequestQueue and pixmapReleaseQueue, in order to
keep track of the sequence numbers associated with the requests.  Adjust
the code using those queues accordingly.
(pixmapRequested), (pixmapReleaseRequested), (themeDaemonStatus): Take
an additional sequenceNumber argument and use it when pushing items into
the queues or when sending out packets to the client.
(clientDataAvailable): Include the sequence number of a request in the
reply packet sent back to the client.
(themeChanged): Send a packet with sequence number 0, as theme changes
are indicated to clients without an explicit request.

* src/theme/duiremotethemedaemon{.cpp,.h,_p.h}: Move include of windows.h
to the top of the file, so that it has a chance to actually work.
(DuiRemoteThemeDaemonPrivate::pixmapRequests): Put the sequence number
of requests as the value in the hash map.
(DuiRemoteThemeDaemonPrivate::nextSequenceNumber): New method which
returns the next available sequence number, incrementing an internal
counter.
(DuiRemoteThemeDaemon::DuiRemoteThemeDaemon): Send a sequence number
along with the RequestRegistrationPacket.
(DuiRemoteThemeDaemon::addDirectoryToPixmapSearchList): Send a sequence
number along with the RequestNewPixmapDirectoryPacket.
(DuiRemoteThemeDaemon::clearPixmapSearchList): Send a sequence number
along with the RequestClearPixmapDirectoriesPacket.
(DuiRemoteThemeDaemonPrivate::waitForServer): Reenable Sleep() call on
Windows platforms, now that the include statement should work.
(DuiRemoteThemeDaemonPrivate::waitForPacket): Instead of waiting on the
type of a packet, wait on a sequence number instead.  This way, requests
and replies can be matched up properly, including ErrorPacket replies.
(DuiRemoteThemeDaemonPrivate::requestPixmap): New internal method to
handle a pixmap request; used by pixmapHandle() and pixmapHandleSync().
Send a sequence number along with the RequestPixmap packet.
(DuiRemoteThemeDaemon::pixmapHandle), (pixmapHandleSync): Implement in
terms of requestPixmap() to get rid of duplicated code.
(DuiRemoteThemeDaemon::releasePixmap): Send a sequence number along with
the ReleasePixmapPacket.  Remove a pending request for the same pixmap
from the pixmapRequests hash map, so that a subsequent request for the
pixmap will result in a new RequestPixmapPacket being sent to the server.
(DuiRemoteThemeDaemonPrivate::pixmapUpdated): Always notify on pixmap
updates, even if a ReleasePixmapPacket has already been queued for the
pixmap in question.  This avoids fragile logic to prevent that scenario,
and the DuiTheme code receiving the notification seems to be able to
handle it just fine.

* src/theme/duitheme.cpp (defaultPixmapCacheId): New internal function
to replace the horrible DEFAULT_PM_CACHE_ID macro.  Code does not become
magically faster just because it is defined as a macro.  Also modify the
string concatenation somewhat to avoid a bit of overhead.
(scalableImageCacheId): Ditto for SCALABLE_IMAGE_CACHE_ID.
(DuiTheme::*), (DuiThemePrivate::*): Adjust references to the old macros
throughout the code to use the new functions instead.

* src/theme/duithemedaemonprotocol.{cpp,h} (Packet): Use m_ prefix for
private members instead of a plain underscore.  Inline trivial accessors
as this is internal API anyway.
(Packet::PacketType): Remove obsolete enumeration values.
(Packet::m_seq): New member to keep track of the sequence number of the
request.
(Packet::sequenceNumber), (setSequenceNumber): New accessor methods.
(PixmapIdentifier::operator!=): New operator to complement operator==().
(operator<<): Write the sequence number to the data stream.
(operator>>): Read the sequence number from the data stream.  Initialize
POD variables used to store output arguments.
(qHash): Instead of converting the pixmap width and height to strings and
concatenating everything, compute separate hashes directly from the data
and bit-twiddle them together.  Hash functions should be fast, and not do
mad things like temporary string allocations.