Changes: make configure script work better
RevBy: TrustMe, Tomas Junnonen

Details: This patch changes the following things:

* make out-of-source builds work.

the problem here was that ther was still a config file
generated in-source. and the move of that file to the
build dir was quite tricky. because the relativity of
pathes in include() statements in .pro files is a bit
strange, we had to have a way to reference the build
dir as an absolute path. this leads to:

* create a .qmake.cache in build dir

this change creates a .qmake.cache dir in the build
tree root. this file is created from the configure
script. and it is read by qmake. then the variable
DUI_BUILD_TREE in there contains the absolute path
to the build tree. now we cna use that variable to
be able to include() files in the build tree. this
leads to:

* generate dui_defines.prf in build dir

before the DUI_PREFIX variable in dui_defines.prf
was hardcoded. and in the end we want the DUI_PREFIX
to be the path that was given as parameter to the
configure script. so we moved the dui_defines.prf
to dui_defines.prf.in and added a marker that is
then replaced by the real install path by the configure
script.

also done:

- fix translation include pathes
- add check for minimum qt version
- change translations.prf to dui_translations.prf

Changes: include dui_defines.prf

RevBy: TrustMe

Details: The original problem this patch fixes is:
when we create the .deb packages with dpkg-buildpackage,
the creation of the i18n packages failed. The reason
was that the created Makefile's for the translations
were missing the install targets for the .qm files.
we tracked this problem down to:

qmake without "-r" recursive parameter can find the
dui_defines.prf in the build dir, although it was referenced
relatively from the source directory.

qmake with "-r" recursive parameter can not find the
dui_defines.prf file in the build dir.

and in debian/rules we were calling "configure" first,
which then calls qmake without "-r", and afterwards
we call qmake with "-r" directly.

so we applied two fixes: we reference the dui_defines.prf
in the build dir with $${DUI_BUILD_TREE}/... and we also
removed the now superfluous call to qmake in debian/rules.