Changes: calculate property names of css attributes just once
RevBy: Kuisma Salonen

Details:
So far while creating style objects the conversion from the
name of an attribute to a correct property name happened
again and again.
This patch changes the behavior to just create the property
name once and forget about the exact name specified in the
css file.