Changes: create style objects lazy
RevBy: DUI-Team Symbio

Details:
So far we create all style objects as soon as a mode
is set on DuiStyleContainer. This includes styles for
both orientations and also styles which will never be
used.
It also simplifies the code generated by duigen and
encapsultes the style creation in DuiStyleContainer.

This patch makes a difference of some130ms when
starting widgetsgallery. One might argue this approach
could slowdown changing the orientation, but if this
really becomes an issue we must fix the way styles
are created (we need to optimize this anyway) and not
delay startup.