Fixes: NB#158975, widgetsgallery crashes when changing language with the "Change Language" dialog
RevBy: John Tapsell, Holger Schröder

Details:

This crash occured because the foreach loop in duiwindow.cpp iterated
over all items on the scene and sent them the language change event,
which could cause deletion of other items handled later in the loop.

I changed the foreach loop in duiwindow.cpp to loop only over the
toplevel items in the scene and call a handler function which
recurses over the children.

And I removed the code from duiwidget.cpp to forward the event to the
child widgets because the handler does this now.