Changes: add some more windows fixes
RevBy: Armin Berres

Details:

Under windows 7 the opengl rendering does not work, so we default
to software rendering under windows.

Simplified duigen creation. the copy step is not needed any more.

Under windows we do not have dbus or gconf. For that reason we
have to save the current language in some other way. Now we do
load and save it in an environment variable.

We also need a hack in the translation catalog loading under
windows right now. otherwise it does not load the translation
catalogs at all.