Changes: add qmake variable for duigen executable, add configure-win32.pl
RevBy: Armin Berres

Details: On windows the duigen executable will be in
duigen/debug/duigen.exe instead of duigen/duigen. So we need a
variable in the qmake .pro files to refer to the executable.
And for that to work we need the DUI_BUILD_TREE variable in
the build directory in the .qmake.cache file. For that i added
the configure-win32.pl file. we do have a dependency on perl
anyway, so i chose to code it in perl.