Changes: adapt doc and examples to maemo-services -> maemo-interfaces renaming
RevBy: Max Waterman

Details: We updated the service framework documentation to reflect
the current state. Also moved the servicefw directory from demos
to examples. Removed the emailservice examples, as they were still
pointing to the eternia demo, which does not exist any more. Adapted
the names of files to the naming conventions we settled upon for
maemo-services, soon to be known as maemo-interfaces.