Changes: DuiSceneManager::orientation{,Angle}() improvements
RevBy: TrustMe
Details: Return values based on DuiSceneManagerPrivate::angle,
instead of DuiSceneManagerPrivate::newAngle. This makes difference
when it comes to querying current orientation (angle) at the very
beginning of the rotation animation. Currently the new angle was
returned, and this patch fixes it to return the old angle from before
the rotation animation. This provides correct orientation info to
DuiStyleSheet (especially when style() is requested in paint(),
drawBackground(), etc.) and in the end improves orientation animation
look and feel.