Fixes: NB#158821 -  dui-servicefwgen should not force shipping of unnecessary include files.
RevBy: Holger

Details:
Missed that I need to include the proxy header file in the cpp. It was previously pulled in via it's header file, which I removed as per bug report, but it is needed for the cpp.
Missed in testing because the default installation root was changed and so I was using a previously installed version rather than the new, improved version.