Changes: Property for not applying style to widgets
RevBy: Torsten Rahn
Details:

Setting a dynamic property named Dui::NoDuiStyle to any QWidget or
QApplication will cause the Qt Maemo 6 style not to style this widget.

The definition of Dui::NoDuiStyle is in duinamespace.h to ensure it can
be used even if the Qt Maemo 6 style is not included in the package.
The preprocessor construct is unfortunately needed to suppress a
"variable not used" warning/error.