Changes: DuiWindow no longer gets in-scene status bar
RevBy: TrustMe
Details: Now only DuiApplicationWindow have the in-scene status bar.
DuiWindow will still be resized and put below the sepatate status bar
top level window like before.

This is because several clients rely on a bare bones DuiWindow which does
not have a scene or a scene manager. Having an in-scene status bar also
brings along with it a scene manager and a scene.