Fixes: NB#161451 - MSceneManager should try to scroll the contents of a page before moving the window
RevBy: Pekka Vuorela, Dominik Kapusta
Details: The scene mananger handles the window relocation requests regarding
input methods, in order to ensure visibility of the focused text edit. This
commit checks whether the window that needs to be relocated is a
MApplicationPage and tries to scroll its contents first. If the required
adjustment could not be achieved by scrolling alone, the window is moved.