Fixes: NB#160444 - Regression: Fennec crashes after opening Add-ons tab
RevBy: Tomas
Details: Remove the scene window from its current scene in
DuiSceneManagerPrivate::PrepareWindowShow(), when it's different than
the scene on which the window is going to be shown. This helps to avoid
a crash when users add a scene window manually to the scene (using e.g.
QGraphicsScene::addItem()), before calling appear().