Changes: Do not block paint events when there are pending requests
RevBy: Mike, John, Tomas

Details:
If the themedaemon has pending events it does not make sense to process
repaint events. This is why normally repaint events when there are
pending events should be ignored.
Right now there is still a possibility that there are bugs in the themedaemon
which could in the worst case completely block redrawing. In the common
case not blocking repainting allows us to immediately see how long themedaemon
needs to load images as they will initially appear as green squares.
This commit should be just a temporary one.