Changes: add command line options to configure to compile without gconf and icu
RevBy: TrustMe

Details:

There is no gconf and no icu for the Windows and the Mac port, which is autodetected.
But even on Linux where gconf and icu should usually be available it is useful
to have options to explicitly disable gconf and icu. Disabling those on Linux
makes it easier to check how the behaviour on Windows and Mac is and whether
it compiles at all without gconf and icu.