Changes: Smart detection of moc path
RevBy: Tomas
Details:
Adapts the mmoc script to better detect the path to the moc
executables. It adds the cases where QTDIR is defined but moc is not installed
in that path (happens on fedora 12 with qt3 installed but not qt3-devel), and
the case where the moc executable is named qt4, which is the case on various
popular Linux distributions.