Changes: define cache directory in pro files
RevBy: TrustMe

Details: The cache directories have been hardcoded in the sources so far,
but as they have a common prefix this should not happen.
The main problem ist that the cache dir for css files have been wrong
and mapplications therefor could not save the binary css files.
The prefix is not configurable via ./configure, I do not consider this
useful.