New: Show StatusBar in fullscreen mode when there's an ongoing call.
RevBy: Daniel d'Andrada
Details: Display status bar regardless of the fullscreen setting in
MApplicationWindow when there's an ongoing call. Connect to the  D-Bus
service that emits a signal when the call is initiated, and update the
relevant flag (callOngoing) in MApplicationWindow. Basing on
callOngoing state, display the status bar in fullscreen mode.