Changes: just change orientation angle when really needed in the timedemo
RevBy: Kuisma Salonen

Details:
So far every benchmark was saving the former orientation angle, rotated the
scene to the angle the benchmark needed and rotated the scene back
to the former angle. If two following benchmarks need the same angle
it does not make sense to always rotate back. Additionally the rotations
somehow messed things up.