Changes: prevent a resource leak
RevBy: Armin Berres, Artem Egorkine
Details: a test method can exit in the middle if a verification fails.
If this happens, any cleanup code placed after that verification doesn't
get executed. That's why it's a good design to do any cleaning up
centrally in the cleanup method - that's what it's for.