Changes: DUI-2683 View Menu and Tapping Back/Close
RevBy: Daniel d'Andrada
Details: As a user, when I have an application menu open and I tap on an escape
button the menu shall close and the escape action shall not be executed. In the
widgets gallery, I can go to the application menu page, open the menu, tap the
escape button and see that the menu has closed and the page didn't go back.