Changes: Add new icon() property on MButton, that takes a QIcon
By: James Ketrenos
RevBy: Tomas Junnonen
Details:
Using MButton::setIcon a caller can now set a QIcon to be used
within the button widget.  When called, the toggledIcon state is set
to the QICon::Selected mode of the QIcon.  A subsequent call to set the
toggledIconID() can set the pixmap data to a state other than what is
in the QIcon.

This is needed for applications which either want to render their own
icon graphics at run-time or want to use an image loader like Qt's
QIcon::fromTheme()