Changes: grabbing all Qt gestures in MWindow.
RevBy: Dominik.
Details: In order to receive gestures in the QGraphicsScene, these
gesture have to be grabbed on the QGraphicsView level. This patch adds
the rest of the Qt gestures to the list of grabbed gestures in MWindow
to reduce amount of code needed from application developers to start
receiving gesture events.