Changes: Use QGraphicsWidget instead of MWidget as the application extension/applet base class
RevBy: Sachin Kundu, Daniel D'Andrada
Details: This patch breaks the libmeegotouchextensions API. It changes the
base class of the extensions to be QGraphicsWidget instead of MWidget so that
in the future extensions can be written without having to use libmeegotouch.
This change has been announced and agreed with the relevant parties.
Approval from asset teams that are using application extensions has been
received and the change has been discussed in DPMT.