Fixes: NB#166495 - No smooth panning for large pages possible
RevBy: Tomas Junnonen
Details: This patch tunes the physics engine to allow better reactions
to very fast movements of the pointer. It postpones the stopping of
panning movement of the viewport to allow recognition to happen first.
It also increases the pan movement threshold to 20px, because when the
user tries to move the viewport really quickly, he quite often crosses
this boundary.