Changes: MPannableViewportLayout - Make the central widget fill the viewport
RevBy: Michal Guminiak
Details: This is a fairly large change in that it means that a widget in a viewport
will be expanded to fill the viewport if it is small and if possible.

This means that it is now possible to place widgets as the bottom of the window.