Changes: use U_ICUDATA_NAME for proper userdata fallback
RevBy: John Tapsell

Details:

If ures_open("ICUDATA-lang", ...) is used, fallbacks to userdata
in usrdt44l/ directories do not work, only original icu data
can be used then. It is better to use
ures_open(U_ICUDATA_NAME "-lang", ...) because this makes the
fallback to user data work.

makes it easier to fix the problem with the language and country
endonym for de_CH.