Changes: Prestarted applications do not allocate glwidget until shown
RevBy: Juha Lintula, Armin Berres, Tomas Junnonen
Details:
Prestarted applications are given qwidgets (instead of qglwidgets) as viewports when they start up.
Viewport is changed on setVisible(true) if the application is not in the prestarted state anymore
and if software acceleration has not been forced.

Initialization and configuration of viewports refactored to separate methods from
setTranslucentBackground.