Changes: fix build failure on Ubuntu (has libicu 4.0.1)
RevBy: John Tapsell

Details:

The qmake check isEqual(ICUVERSION, 4.2.1) was good enough for maemo
because there we have eihter 4.2.1 or 4.4.1, not other versions, but
it failed on Ubuntu because this had 4.0.1.  Using
greaterThan(ICUVERSION, 4.4) instead works.  This is a string
comparision, not a numeric one, but it works good enough for our
purpose.