Fixes: NB#171579 - "widgetsgallery -quitimmediately" does not terminate anymore
RevBy: John Tapsell, Esko Oramaa, Ville Voutilainen

Details:
You are not allowed to read from a QIODevice with operator<< without beeing sure
that the bytes you want to read are actually available. This commit adds the
necessary checks and waits for data if needed.
It was just pure luck that the current solution was working at all. Not all request
were succeeding, we just did not notice the failing ones so far.