Fixes: NB#173036 - Adding rows to model updates all visible items in MListView
RevBy: Sergiy Dubovik

Details:
The updateGeometry sends an event to the layout to re-request the sizeHint of the
list, but the list is resized and relayout-ed after 1 frame is drawn empty, this causes
the list to flicker. The following fixes the behavior, by setting the flag that the
visible items array should be cleared on next relayout.