Changes: While the application menu is open, disable unchecked toolbar buttons.
RevBy: TrustMe
Details: To comply with interation design of the application menu.
While the menu is open the toolbar should not react to any user input,
but simply disabling it (setEnabled(false)) would have the side
effect of unchecking the selected item in the tab view of the toolbar.

To avoid this side effect this patch disables all toolbar widgets but
the one that is checked. Tapping the checked item in the tab view will
have no effect anyway, so there's no harm in keeping it enabled.