Changes: make ut_mapplicationinstancemanager work with mmoc
RevBy: Max Waterman
Details: here the problem was that some header files in this
test do contain M_MODEL_PROPERTY lines, but only qt moc
was used to build this test instead of mmoc. and so some
function declarations were not created. when i switched to
mmoc, the function declarations were created, and as the
cpp files were missing from the .pro file, the linker
was complaining about missing symbols.