Changes: implement city and country list for libmeegotouch
RevBy: TrustMe

Details: these classes can be used to query for a list
of cities and countries. every city holds information
about its associated timezone.

this can be used in the startup screen to select the
location of the user, and it can also be used in the
clock application to select different time zones for
a world clock etc.