Fixes: NB#173416 - setting gconf value for /meegotouch/theme/name only works once
RevBy: Tomas Junnonen
Details: If an invalid theme gconf value is inputted the mthemedaemon does not forcefully change
it to valid one, but silently falls back to use the default theme. Implemented functionality for
changing theme even though previous theme change is still in process. The new theme is applied
after the previous change is completed.