Changes: Fixed autohiding components disappearing, appearing, then immediately disappearing again
RevBy: TrustMe
Details: Don't call appear on scene windows of the MApplicationWindow that are in the process
of disappearing in the mouse release event. The release is processed after enabling the autohide
via a UI component, or transitioning to a subview with the autohide mode already set, which
was always causing the multiple transitions for the scene window.