Changes: document that formatDateTime() does not uppercase the first letter.
RevBy: Torsten Rahn, John Tapsell

Details:

MLocale’s formatDateTime() methods do not automatically uppercase the
first letter of the result because they cannot know in which context
the result will be used and depending on the context uppercasing maybe
correct or wrong.

Therefore it is left to the application to do the uppercasing if
needed.

Document this and give an example how an application could do the uppercasing
and put the sample code also in languagepage.cpp in widgetsgallery.