Fixes: NB#176455 - MClassFactory trashes debug output
RevBy: John, Michal

Details:
Currently enabling debug output is completely impossible as
MClassFactory prints tons of messages. The debug output is
disabled now. The output can be usefull sometimes but in most
of the cases it is just telling us about Qt classes or pure
virtual classes or classes which have no widget creator for
a good reason.
Additionally we do not try to get the assembly name if we
already now there is no matching assembly.
QMaps have been replaced with QHash as the past showed they
are faster in general -- in this very case I could just measure
a small difference, if there was one at all.
As the C++ standard says this: "An expression of arithmetic or
enumeration type can be converted to an enumeration type
explicitly. The value is unchanged if it is in the range of
enumeration values of the enumeration type; otherwise the
resulting enumeration value is unspecified." I also added an
explicit enum value for not found assemblies.