Changes: inline call to MStyleContainer::operator->()
RevBy: Peter Penz

Details:
operator->() called on all types of style containers is one of the most
executed methods of every Meego Touch application. By inlining this call
we can save quite some method calls.
As inlining a method is a binary incompatible change a dummy method is
introduced which forces gcc to also create a non-inlined version of the
method. When we are allowed to do the next binary incompatible changes
it should be removed.
See also NB#174717.