Fixes: NB#175663 - system dialog cannot be minimised
RevBy: Marcin Miklas, Tomas Junnonen
Details: Introducing system property for MDialog. It enables
creating System Dialogs that are modeless (when modal property
is set to false). The difference between System Dialog and System
Modal Dialog is that the former is equipped with the home button
that allows skipping the dialog temporarily and later invoking
it from the task switcher.