Changes: delay loading of filesystem cache in the themedaemon
RevBy: Peter Penz

Details:
Second try.
The themedaemon has a filesystem cache to cache pixmaps.
At startup or when changing the theme right now the whole cache is
read at once. If an element is not found in the cache it is loaded
delayed as soon as the device is idle.
With this patch the file system chache is also loaded delayed to
speedup startup and theme changing times.