Fixes: NB#178001 - using MList::scrollTo() makes the list behave strange
RevBy: Jörgen

Details: The list postponed the geometry update event to next event loop,
so if scroll-to was invoked before the geometry being ready, it would result
in pannable viewport not yet being ready for scrolling and wrong target
position for scrolling being calculated. The fix ensures that if pannable
viewport range is changed, then the target position for scrolling is recalculated.