Changes: introduce packet to communicate most used pixmaps
RevBy: TrustMe

Changes:
The themedaemon internally keeps a list of most used pixmaps which are
constantly loaded. Right now this list is not communicated to the clients.
With the newly introduced packet the clients can be informed about
the most used pixmaps and their update.
Additionally some streaming operators have been introduced for convenience
and two bugs when reading QStrings and QStringLists have been fixed.