Changes: block signals from socket while reading packet
RevBy: TrustMe

Details:
After debugging communication problems between clients and the themedaemon
for hours I noticed that we start reading a second packet when the former
one has not yet been completely processed. This did not happen so far
as the packets are very small. I am about to introduce a huger packet which
triggers the problem.