Changes: themedaemon sends most used pixmap handles to the clients
RevBy: Peter Penz

Details:
The themedaemon keeps internally a list of most used pixmaps which
are always hold in memory. When a client requests a pixmap from this
list the pixmap can be returned faster. Ideally the client would
not have to query the themedaemon at all but just know about the
handles if the most used pixmaps. With this commit the themedaemon
sends the handles of the most used pixmaps to the clients when they
register and keeps them up to date if the list of most used pixmaps
changes.
If a client uses one of these pixmaps it notifies the themedaemon so
it can keep statistics about most used pixmaps up to date.