Changes: Block the painting in MWindow until the layouting has been finished
RevBy: Armin Berres

Details: Because of the recent optimizations of the theme daemon, pending requests for images are very rare. So the approach to delay the showing of the window until all pending requests are done, is not sufficient anymore to prevent a flickering of applications during startup. The result is that the raw application window can be seen for a fraction of a second during startup. This patch fixes the issue by blocking the painting until QEvent::WindowActivate has been received. Additionally the if-else-if cascading in MWindow::event() has been replaced by a switch-command.