Changes: Improvements to chaining example.
RevBy: TrustMe

Details:
Added an INSTALL.txt file to examples/chaining - list of instructions to show one method of installing this on a device.
Made the name of the applications different at the start of the name, rather than the end - the launcher cutts it off so you can't tell between them.
Made the button labels a little more obvious so you can tell between the window which is launched through chaining and the one that isn't.
Changed chaining example app to be valid - chainingexampleapp2 was opening a window when it was launched which caused a few problems. I think such behaviour is not typical, so removed it.