Changes: revive themedaemon stress test
RevBy: TrustMe

Details:
The stress test did not work anymore. I updated the used themes and
ported it to use MRemoteThemeDaemon instead of an own implementation.

Not all problems are fixed yet, when multiple clients are running
sometimes the themedaemon and the clients get slightly out of sync
for a short timeframe.