Changes: messageid-check script: add exceptions for qtn_clk_{city,country,region}
RevBy: TrustMe

Details:

meegotouch-cities-template does not create an engineering English
package at the moment. I am not sure whether it should or not. If it
did, would the engineering English package need to be customized as
well?

At the moment I mark these as used as an exception because I happen to
know that they are used.