Changes: Fix installation of pkgconfig .pc files
RevBy: Mike Fabian
Details: in an out-of-source build, as it is done when
building .deb packages, we must make sure that the
destination directory exists. that is what the mkdir is for.
and then we create the generated files in the outdir, and
the content is from the reldir. works for me for a debian
build.