Fixes: NB#182098 - List with separators on renders incorrectly
RevBy: Sergiy

Details: The list view private class was updating separator
size just after the separtor style was set, and then once more after the style
was applied. The fix ensures that the separator size is updated only
once on either geometry or style update, and the correct overloaded function
is called.