Changes: themedaemon extract ids from all svgs
RevBy: Janne Heikkilä, Peter Penz, Tomas Junnonen

Details:
When the themedaemon is supposed to return an pixmap for a special id it could
theoretically be part of any available svg. In the worst case the themedaemon
has to iterate over all svgs to find an id. To speed this up so called .ids files
could be used to specify which ids are contained in an svg. If an id from
an .ids file has been requested the lookup has been fast. Ff not the
themedaemon had to to the iteration. Quite some svg files did not have .ids
files.
To solve this problem the .ids files are obsolet with this commit. The
themedaemon extracts the ids contained in all svgs now at startup and caches
the result. When an id is requested now the themedaemon can use a lookup table
to directly determine the needed svg file.
As extracting the ids from the svgs is not the cheapest opertion the ids
are saved in cache file which are updated whenever the svg changes.