Changes: reduce themedaemon memory usage
RevBy: Peter

Details:
When loading a png/jpg image or loading an icon the themedaemon
was always keeping a QImage/QSvgRenderer in memory as long as a
client requested the theme element. Not the QImage is freed imediately
and the QSvgRenderer will be freed when it is unused for three seconds.
Memory usage us reduced from a bit more than 6Mb to a bit more than
4Mb.