Changes: store rendered svgs in binary form
RevBy: Jakub Pavelek, Peter Penz

Details:
So far only the most used pixmaps have been stored in a binary form.
Now every image which is created from a scg file is stored in its
rendered binary form. This cache will be saved on disk and the svg
file does not need to be opened again when the id is requested the
next time.
This gives us reliable response times in the themedaemon and reduces
memory usage/fragmentation.