Fixes: NB#178931 - System modal dialog cannot be invoked twice
RevBy: Michał
Details: Guard MSceneWindowPrivate::sceneManager pointer. When the
system modal dialog is closed and it's window is deleted (together with
the scene manager), the pointer is zeroed and upon next call to appear
it will be recognised as invalid pointer and it won't be dereferenced,
especially in MSceneManagerPrivate::addSceneWindow().