Changes: make overriding ICU resources work for non-MApplications as well
RevBy: John Tapsell

Details:

NB#169260 has been fixed a while ago but comment#30 claimed that the
problem still exists, so I investigated again. It works fine for me in
the clock application, so I think the bug is fixed, but while
investigating I found that it does not work in applications, which are
not MApplications but only QCoreApplications.

This commit fixes this additional problem.