Changes: MLayout - Fallback to orientation of activeWindow if not attached to a window
RevBy: TrustMe
Details: When a layout is attached to an item or placed inside another layout, we don't get notified.
So we need to guess at what the correct orientation is before hand.  We do this by taking the orientation
of the currently active window if we are not attached to any window.

This fixes several unit test failures.