Changes: load pixmaps based on priorities
RevBy: Peter

Details:
So far the themedaemon loads pixmaps based on a first come first
serve base. When a bunch of applications request of lot of pixmaps,
this happens e.g. at startup or during a theme change it is pure
pure chance which item is loaded first. In the worst case an
invisible background application receives its pixmaps before the
forground application.
With this commit all pixmap requests have a priority. The highest
priority have currently visible applications, the lowest one pre-
started ones. Invisible not prestarted applications come in
between. Additionally via a conf file a set of applications be be
specified which have a fixed priority. Right now the forground
app has a priority of 100, sysuid 90, duihome = 80,
meego-im-uiserver 70. This means they will be updated before the
invisible applications which have a priority of 0. Change
themedaemonpriorities.conf if other values are needed.