Fixes: NB#176079 -  MWidgetController does not connect to MWidgetModel::modified by default
RevBy: Tomas Junnonen
Details: Because we cannot call a virtual function from the constructor, we need to delay the
call of setupModel().  However we do not need to delay this call if setModel is not called from
the constructor. And we can at least connect the dataChanged signal immediately.

Includes unit test and doxygen improvements.