Fixes: Swipe gesture will now snap to 0,90,180 and 270 degrees.
RevBy: Tomas Junnonen
Details: Meegotouch swipe gesture recognizer is reusing Qt's QSwipeGesture class
which is delivered to the users of the swipe gesture. In this class, the
swipe direction is dynamically calculated from angle attribute. The algorithm
used there makes it almost impossible to recognize the gesture fully in
vertical or horizontal direction. To help that, we are introducing snapping
to the right angles. If the user will swipe in an angle that is close enough (by 22.5 degrees),
the recognizer will change the angle to the right angle.