Fixes: NB#180607  -  Mouse events and other gestures active during pinch gesture
RevBy: Tomas Junnonen
Details: Our gesture recognizers were not checking if the count of the fingers
on the screen matches the description of the gesture. This is particularly
important to the widgets that are registered to single and multitouch gestures
like pan and pinch and want the single finger gestures be canceled when pinching
is starting.

Please note, that by default the gesture recognizer does not make your widget
accept the touch events. In order to start receiving touch events you need to
call widget->setAcceptTouchEvents(true).