Changes: do not needlessly initialize viewport when setting transparency
RevBy: Tomas Junnonen

Details:
So far MWindow::setTranslucentBackground() was recreating the
opengl viewport in non-software mode by calling initGLViewport().
It is called by MWindowPrivate::init() and may be called one more
time by apps before showing the window.
As initGLViewport() is also called by MWindow::setVisible() when
the window is appearing the viewport has been created at least
two times and destroyed one time.
This commit delays the initialization of the gl viewport if it
is not created yet. It will then be created by the
MWindow::setVisible() call.
Another side effect is that a MWindow does not hog memory until
it is finally visible.