Changes: make initialization of MGLES2Renderer partially lazy
RevBy: Peter Penz

Details:
Everytime when creating an MWindow MGLES2Renderer was doing some
initalizations. The most costly part was compiling and linking
a shader. As 99% of the applications do not use MGLES2Renderer
the initialization is now delayed until someone uses the shader.

The speedup we see is up to 5%-10% of the startup of
widgetsgallery.

The only thing I found using shaders is MVideoWidget. It still
works as expected after this commit.