Fixes: NB#181896 - blank window with title "Menu" becomes visible when using invoker
Details: When MComponentCache is populated, it instantiates
MApplicationWindow for applications to use. If an application
instantiates another MApplicationWindow instead of using the instance
from the cache, the instance in the cache was still the active
window. This caused unexpected behavior like application pages to
appear to a wrong window by default. This commit fixes the issue by
delaying registration (and thus activation) of the window in the cache
until it is taken from there.