Fixes: NB#179986 - MImageWidget does not take aspect ratio of images (logical image ID's) into account when specifying image sizes.
RevBy: Marcin, Jörgen

By: Sergiy Dubovik, Stanislav Ionascu

Details: In case if the size of the theme-based image is not specified,
the preferred size of the image widget is used to request the pixmap.
If the requested size is explicitly specified, then that size is used.
Also moved the image-id and the requested image size into MImageWidgetModel
for consistency.