Fixes: NB#164207 -  long press in MLabel
RevBy: Stanislav

Details: This patch fixes MLabel long press functionality.
It changes the MTapAndHoldRecognizer as suggested by Qt. The main
part of the workaround is that the recognizer object is created only
for the graphics view widget, not for the graphics items. This fixes
the bug with gesture event delivery when multiple gesture objects
were delivered to graphics items.

In our case this means that the gesture recognizers will now need to work
on QMouseEvents instead of QGraphicsSceneMouseEvents.